refactor: extract liblog + terminal helpers, move inline tests to src/main_tests
main.rs 5172 -> 4370 (-802). No behavioral change.
