+++ src/lib.rs
@@ -13,6 +13,7 @@ pub mod components;
 pub mod cover;
 pub mod gradient;
 pub mod httpcache;
+pub mod liblog;
 pub mod lyrics;
 pub mod reactive;
 pub mod theme;
@@ -24,6 +25,8 @@ pub mod audio;
 pub mod config;
 #[cfg(feature = "streaming")]
 pub mod engine;
+#[cfg(feature = "streaming")]
+pub mod term;
 #[cfg(feature = "streaming")]
 pub mod webapi;
