+++ src/main.rs
@@ -5,27 +5,22 @@
 //! Multi-section library (playlists / liked / albums / artists), shuffle, repeat,
 //! and a live queue view.
 
-use std::io::{self, Stdout};
+use std::io;
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
 
 use anyhow::{Context, Result};
 use crossterm::event::{
-    self, Event, KeyCode, KeyEventKind, KeyModifiers, KeyboardEnhancementFlags, MediaKeyCode,
-    MouseButton, MouseEventKind, PopKeyboardEnhancementFlags, PushKeyboardEnhancementFlags,
+    self, Event, KeyCode, KeyEventKind, KeyModifiers, MediaKeyCode, MouseButton, MouseEventKind,
 };
 use crossterm::execute;
-use crossterm::terminal::{
-    disable_raw_mode, enable_raw_mode, BeginSynchronizedUpdate, EndSynchronizedUpdate,
-    EnterAlternateScreen, LeaveAlternateScreen,
-};
-use ratatui::backend::CrosstermBackend;
+use crossterm::terminal::{BeginSynchronizedUpdate, EndSynchronizedUpdate};
 use ratatui::buffer::CellDiffOption;
 use ratatui::layout::{Alignment, Constraint, Layout, Margin, Rect};
 use ratatui::style::{Modifier, Style};
 use ratatui::text::{Line, Span};
 use ratatui::widgets::{Block, Clear, Paragraph};
-use ratatui::{Frame, Terminal};
+use ratatui::Frame;
 use ratatui_image::picker::Picker;
 
 use myx::anim::ThemeFade;
@@ -34,8 +29,10 @@ use myx::components::{gradient_line, gradient_progress, left_bar_block};
 use myx::cover::Cover;
 use myx::engine::{self, Engine, EngineEvent};
 use myx::gradient::{self};
+use myx::liblog::{install_librespot_log, liblog};
 use myx::lyrics::parse::parse_lrc;
 use myx::reactive::derive_theme;
+use myx::term::{acquire_single_instance_lock, init_terminal, restore_terminal, Term};
 use myx::theme::{Theme, TOKYONIGHT};
 use myx::util::{center_v, fmt_ms, track_id_from_uri, truncate, uri_to_url, urlencode, vol_u16};
 use myx::webapi::WebApi;
@@ -44,7 +41,6 @@ use souvlaki::{
     SeekDirection,
 };
 
-type Term = Terminal<CrosstermBackend<Stdout>>;
 const FADE_MS: u64 = 1500;
 
 /// How far one Shift+arrow press moves the playhead.
@@ -2537,80 +2533,6 @@ fn apply_meta(
 
 // ------------------------------------------------------------------ web api
 
-/// Temporary-but-useful diagnostics for startup/library failures. Kept out of
-/// the TUI because alternate-screen rendering hides stderr.
-/// Forwards librespot's own `log` output into `myx.log`. Its Connect
-/// diagnostics are the only account of why the device went inactive, and
-/// without a logger installed they go nowhere.
-struct LibrespotLog;
-
-impl log::Log for LibrespotLog {
-    fn enabled(&self, _: &log::Metadata) -> bool {
-        true
-    }
-    fn log(&self, record: &log::Record) {
-        liblog(format!(
-            "{} {}: {}",
-            record.level(),
-            record.target(),
-            record.args()
-        ));
-    }
-    fn flush(&self) {}
-}
-
-/// Any value of `MYX_LOG` turns logging on; the value only picks how loud
-/// librespot is. `debug`/`trace` open it up, `warn` quiets it back down.
-fn install_librespot_log() {
-    let Ok(level) = std::env::var("MYX_LOG") else {
-        return;
-    };
-    let filter = match level.to_ascii_lowercase().as_str() {
-        "trace" => log::LevelFilter::Trace,
-        "debug" => log::LevelFilter::Debug,
-        "warn" => log::LevelFilter::Warn,
-        // Connect names its own faults at info ("device became inactive");
-        // warn shows only the fallout.
-        _ => log::LevelFilter::Info,
-    };
-    if log::set_boxed_logger(Box::new(LibrespotLog)).is_ok() {
-        log::set_max_level(filter);
-    }
-}
-
-/// Optional debug log — silent unless `MYX_LOG` is set. Writes to
-/// ~/.cache/myx/myx.log (user-owned dir 0700, file 0600) instead of a
-/// world-writable fixed /tmp path (audit H5).
-fn liblog(msg: impl AsRef<str>) {
-    use std::io::Write;
-    if std::env::var_os("MYX_LOG").is_none() {
-        return;
-    }
-    let Some(home) = myx::home_dir() else { return };
-    let dir = home.join(".cache/myx");
-    if std::fs::create_dir_all(&dir).is_ok() {
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let _ = std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(0o700));
-        }
-    }
-    let mut opts = std::fs::OpenOptions::new();
-    opts.create(true).append(true);
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::OpenOptionsExt;
-        opts.mode(0o600);
-    }
-    if let Ok(mut f) = opts.open(dir.join("myx.log")) {
-        let ts = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs_f64())
-            .unwrap_or(0.0);
-        let _ = writeln!(f, "{ts:.3} {}", msg.as_ref());
-    }
-}
-
 fn token_of(webapi: &Arc<Mutex<WebApi>>) -> Option<String> {
     // Refresh when the token is expiring so long sessions don't silently go
     // read-only (audit H1). Only holds the lock across the network call in the
@@ -4383,33 +4305,6 @@ fn http_client() -> reqwest::blocking::Client {
 
 // ------------------------------------------------------------------ terminal
 
-/// Hold an exclusive lock so only one myx runs at a time. Returns the lock file
-/// (kept alive for the process lifetime; the OS releases it on exit, even a crash).
-fn acquire_single_instance_lock() -> std::fs::File {
-    use fs2::FileExt;
-    let path = myx::home_dir()
-        .map(|h| h.join(".cache/myx/lock"))
-        .unwrap_or_else(|| std::path::PathBuf::from("/tmp/myx.lock"));
-    if let Some(dir) = path.parent() {
-        let _ = std::fs::create_dir_all(dir);
-    }
-    let file = std::fs::OpenOptions::new()
-        .create(true)
-        .truncate(false)
-        .write(true)
-        .open(&path)
-        .expect("open lock file");
-    if file.try_lock_exclusive().is_err() {
-        eprintln!("myx is already running (another instance holds the lock).");
-        eprintln!(
-            "Close it first, or remove {} if it's stale.",
-            path.display()
-        );
-        std::process::exit(1);
-    }
-    file
-}
-
 const SPINNER: [&str; 8] = ["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧"];
 
 /// Run `task`, drawing the startup screen until it finishes.
@@ -4468,705 +4363,8 @@ fn render_loading(f: &mut Frame, label: &str, frame: usize) {
     );
 }
 
-fn init_terminal() -> Result<Term> {
-    // Restore the terminal on panic so a crash doesn't strand the user in a
-    // raw-mode / alt-screen shell (audit H6). Runs before the default hook (and
-    // before the abort under panic=abort).
-    let default_hook = std::panic::take_hook();
-    std::panic::set_hook(Box::new(move |info| {
-        let mut out = io::stdout();
-        let _ = execute!(
-            out,
-            crossterm::event::DisableMouseCapture,
-            crossterm::event::DisableFocusChange,
-            LeaveAlternateScreen,
-            crossterm::cursor::Show
-        );
-        let _ = disable_raw_mode();
-        default_hook(info);
-    }));
-
-    enable_raw_mode()?;
-    let mut stdout = io::stdout();
-    execute!(
-        stdout,
-        EnterAlternateScreen,
-        crossterm::event::EnableMouseCapture,
-        // Notices a return to this tmux window, when art must be re-sent.
-        crossterm::event::EnableFocusChange
-    )?;
-    // Media key support requires keyboard enhancement (Windows Terminal, kitty, etc.).
-    // Silently skip on terminals that don't support it (legacy Windows console).
-    let _ = execute!(
-        stdout,
-        PushKeyboardEnhancementFlags(KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES)
-    );
-    Ok(Terminal::new(CrosstermBackend::new(stdout))?)
-}
-
-fn restore_terminal(terminal: &mut Term) -> Result<()> {
-    disable_raw_mode()?;
-    execute!(
-        terminal.backend_mut(),
-        crossterm::event::DisableMouseCapture,
-        crossterm::event::DisableFocusChange,
-        LeaveAlternateScreen
-    )?;
-    let _ = execute!(terminal.backend_mut(), PopKeyboardEnhancementFlags);
-    terminal.show_cursor()?;
-    Ok(())
-}
-
 // ------------------------------------------------------------------ tests
 
 #[cfg(test)]
-mod playlist_tests {
-    use super::*;
-    use serde_json::json;
-
-    fn ctx_row() -> LibItem {
-        LibItem::ctx(
-            "Chill Vibes".into(),
-            "you · 142".into(),
-            "spotify:playlist:1".into(),
-        )
-    }
-
-    // ------------------------------------------------------ startup ordering
-
-    #[test]
-    fn authentication_finishes_before_terminal_initialization() {
-        let order = std::cell::RefCell::new(Vec::new());
-
-        let result = auth_then_terminal(
-            || {
-                order.borrow_mut().push("auth");
-                Ok("authenticated")
-            },
-            || {
-                order.borrow_mut().push("terminal");
-                Ok("terminal")
-            },
-        )
-        .expect("startup succeeds");
-
-        assert_eq!(result, ("authenticated", "terminal"));
-        assert_eq!(*order.borrow(), ["auth", "terminal"]);
-    }
-
-    #[test]
-    fn authentication_failure_never_initializes_terminal() {
-        let terminal_initialized = std::cell::Cell::new(false);
-
-        let result: Result<((), ())> = auth_then_terminal(
-            || anyhow::bail!("cached refresh token was rejected"),
-            || {
-                terminal_initialized.set(true);
-                Ok(())
-            },
-        );
-
-        assert!(result.is_err());
-        assert!(!terminal_initialized.get());
-    }
-
-    // ------------------------------------------------ optional integrations
-
-    #[test]
-    fn optional_integration_keeps_successful_service() {
-        assert_eq!(
-            optional_integration(true, || Ok::<_, ()>("media")),
-            Some("media")
-        );
-    }
-
-    #[test]
-    fn optional_integration_degrades_on_initialization_failure() {
-        assert_eq!(
-            optional_integration(true, || Err::<(), _>("no session bus")),
-            None
-        );
-    }
-
-    #[test]
-    fn optional_integration_skips_init_when_platform_is_unavailable() {
-        let called = std::cell::Cell::new(false);
-        let service = optional_integration(false, || {
-            called.set(true);
-            Ok::<_, ()>("media")
-        });
-        assert_eq!(service, None);
-        assert!(!called.get());
-    }
-
-    #[test]
-    fn disconnected_media_channel_disables_future_receives() {
-        let mut open = true;
-        let event: Result<(), flume::RecvError> = Err(flume::RecvError::Disconnected);
-        assert_eq!(consume_media_event(event, &mut open), None);
-        assert!(!open);
-    }
-
-    #[test]
-    fn active_scrub_rejects_stale_engine_position() {
-        assert!(!should_apply_engine_position(true, Some(42_000)));
-        assert!(should_apply_engine_position(true, None));
-        assert!(should_apply_engine_position(false, Some(42_000)));
-    }
-
-    // -------------------------------------------------------- context_target
-
-    #[test]
-    fn context_target_accepts_context_rows() {
-        let (uri, name) = context_target(&ctx_row()).expect("playlist is a context");
-        assert_eq!(uri, "spotify:playlist:1");
-        assert_eq!(name, "Chill Vibes");
-    }
-
-    #[test]
-    fn context_target_accepts_synthesized_play_row() {
-        // "▶︎ Play X" rows carry the context URI, so P works inside a drill-in.
-        let row = LibItem::play("▶︎ Play Chill Vibes".into(), "spotify:playlist:1".into());
-        assert_eq!(
-            context_target(&row).map(|(u, _)| u),
-            Some("spotify:playlist:1".to_string())
-        );
-    }
-
-    #[test]
-    fn context_target_rejects_tracks_and_headers() {
-        let track = LibItem::track("Song".into(), "Artist".into(), "spotify:track:9".into());
-        assert!(context_target(&track).is_none());
-        assert!(context_target(&LibItem::header("Songs")).is_none());
-    }
-
-    // --------------------------------------------------- parse_playlist_track
-
-    #[test]
-    fn parses_an_items_entry() {
-        // The shape /playlists/{id}/items actually serves today.
-        let it = json!({"added_at": "2024-01-01T00:00:00Z", "is_local": false, "item": {
-            "name": "Coffee",
-            "uri": "spotify:track:429NtPmr12aypzFH3FkN9l",
-            "type": "track",
-            "artists": [{"name": "beabadoobee"}]
-        }});
-        let li = parse_playlist_track(&it).expect("valid item");
-        assert_eq!(li.name, "Coffee");
-        assert_eq!(li.subtitle, "beabadoobee");
-        assert_eq!(li.uri, "spotify:track:429NtPmr12aypzFH3FkN9l");
-        assert!(li.is_track);
-    }
-
-    #[test]
-    fn still_parses_legacy_track_entry() {
-        // Older /tracks shape, kept working through the API migration.
-        let it = json!({"track": {
-            "name": "Sailor Song",
-            "uri": "spotify:track:abc",
-            "artists": [{"name": "Gigi Perez"}]
-        }});
-        let li = parse_playlist_track(&it).expect("valid track");
-        assert_eq!(li.name, "Sailor Song");
-        assert_eq!(li.subtitle, "Gigi Perez");
-    }
-
-    #[test]
-    fn skips_null_entries() {
-        // Real playlists contain these for items pulled from the catalogue.
-        // Must yield None (skipped) rather than panic or abort the page.
-        assert!(parse_playlist_track(&json!({ "item": null })).is_none());
-        assert!(parse_playlist_track(&json!({ "track": null })).is_none());
-        assert!(parse_playlist_track(&json!({})).is_none());
-    }
-
-    #[test]
-    fn skips_entry_without_uri() {
-        let it = json!({"item": {"name": "No URI", "artists": [{"name": "X"}]}});
-        assert!(parse_playlist_track(&it).is_none());
-    }
-
-    #[test]
-    fn missing_artists_yields_empty_artist_not_skip() {
-        let it = json!({"item": {"name": "Untitled", "uri": "spotify:track:z"}});
-        let li = parse_playlist_track(&it).expect("still playable without artists");
-        assert_eq!(li.subtitle, "");
-    }
-
-    #[test]
-    fn total_prefers_items_over_legacy_tracks() {
-        // Live /me/playlists shape: `items.total`, no `tracks` object at all.
-        assert_eq!(
-            playlist_total(&json!({"items": {"href": "…", "total": 155}})),
-            Some(155)
-        );
-        assert_eq!(playlist_total(&json!({"tracks": {"total": 42}})), Some(42));
-        assert_eq!(
-            playlist_total(&json!({"items": {"total": 7}, "tracks": {"total": 9}})),
-            Some(7)
-        );
-        assert_eq!(playlist_total(&json!({"name": "no counts"})), None);
-    }
-
-    #[test]
-    fn admits_local_files_and_episodes() {
-        // Documents current behaviour: both parse as ordinary tracks.
-        let local = json!({"is_local": true, "item": {
-            "name": "Demo.mp3", "uri": "spotify:local:::Demo:180", "artists": [{"name": "Me"}]
-        }});
-        assert!(parse_playlist_track(&local).is_some());
-
-        let episode = json!({"item": {
-            "name": "Ep 12", "uri": "spotify:episode:e1", "type": "episode", "artists": []
-        }});
-        let li = parse_playlist_track(&episode).expect("episodes are admitted today");
-        assert_eq!(li.subtitle, "");
-    }
-
-    // ------------------------------------------------------ playlist_subtitle
-
-    #[test]
-    fn subtitle_puts_count_before_owner() {
-        // Count leads so it survives tail-first truncation in a narrow pane.
-        assert_eq!(
-            playlist_subtitle("ImLordVisssh", Some(155)),
-            "155 · ImLordVisssh"
-        );
-        assert_eq!(playlist_subtitle("you", None), "you");
-        assert_eq!(playlist_subtitle("", Some(12)), "12");
-        assert_eq!(playlist_subtitle("", None), "");
-    }
-
-    // -------------------------------------------------------- meta_is_current
-
-    #[test]
-    fn stale_metadata_replies_are_dropped() {
-        let a = "spotify:track:AAA";
-        let b = "spotify:track:BBB";
-        // Waiting on B: B's reply applies, A's late reply does not.
-        assert!(meta_is_current(Some(b), b));
-        assert!(!meta_is_current(Some(b), a));
-        // Nothing outstanding -> accept (the guard only drops provable mismatches).
-        assert!(meta_is_current(None, a));
-    }
-
-    // ------------------------------------------------------------ enter_label
-
-    #[test]
-    fn enter_label_matches_context_target() {
-        let track = LibItem::track("Song".into(), "Artist".into(), "spotify:track:9".into());
-        assert_eq!(enter_label(Some(&ctx_row())), "open");
-        assert_eq!(enter_label(Some(&track)), "select");
-        assert_eq!(enter_label(Some(&LibItem::header("Songs"))), "select");
-        assert_eq!(enter_label(None), "select");
-
-        // The invariant the footer relies on: Enter says "open" for exactly
-        // the rows P can play.
-        for row in [ctx_row(), track, LibItem::header("Songs")] {
-            let opens = enter_label(Some(&row)) == "open";
-            assert_eq!(opens, context_target(&row).is_some() && !row.is_play);
-        }
-    }
-}
-
-#[cfg(test)]
-mod nav_tests {
-    use super::*;
-
-    // -------------------------------------------------------- scroll_offset
-
-    /// Walk the cursor from `from` to `to` one row at a time, as the arrow keys
-    /// do, threading the offset through — the sticky viewport only makes sense
-    /// as a sequence. Returns the offset the walk ends on.
-    fn walk(mut offset: usize, from: usize, to: usize, cap: usize, total: usize) -> usize {
-        let step = if to >= from { 1isize } else { -1 };
-        let mut sel = from as isize;
-        while sel != to as isize {
-            sel += step;
-            offset = scroll_offset(offset, sel as usize, cap, total, 3);
-        }
-        offset
-    }
-
-    #[test]
-    fn short_lists_never_scroll() {
-        assert_eq!(scroll_offset(0, 4, 10, 5, 3), 0);
-        assert_eq!(scroll_offset(0, 9, 10, 10, 3), 0, "total == cap still fits");
-    }
-
-    #[test]
-    fn top_of_a_long_list_stays_put() {
-        // Nothing to reveal above row 0, so no margin is owed there.
-        for sel in 0..=4 {
-            assert_eq!(scroll_offset(0, sel, 8, 50, 3), 0, "sel={sel}");
-        }
-    }
-
-    #[test]
-    fn scrolling_down_keeps_the_margin_below_the_cursor() {
-        // cap=8, margin=3: the cursor may reach row 4, then the list moves.
-        let offset = walk(0, 0, 20, 8, 50);
-        assert_eq!(20 - offset, 4, "3 rows stay visible below the cursor");
-    }
-
-    #[test]
-    fn scrolling_back_up_keeps_the_margin_above_the_cursor() {
-        let offset = walk(0, 0, 30, 8, 50);
-        let offset = walk(offset, 30, 10, 8, 50);
-        assert_eq!(10 - offset, 3, "3 rows stay visible above the cursor");
-    }
-
-    #[test]
-    fn the_cursor_moves_inside_the_viewport_before_the_list_does() {
-        // The whole point of the sticky offset: a short move well inside the
-        // window must not scroll at all.
-        let offset = walk(0, 0, 30, 20, 100);
-        assert_eq!(scroll_offset(offset, 20, 20, 100, 3), offset);
-    }
-
-    #[test]
-    fn the_end_of_the_list_is_reachable() {
-        // The last row must still be selectable — the bottom margin cannot push
-        // the viewport past the end.
-        assert_eq!(walk(0, 0, 49, 8, 50), 42, "clamped to total - cap");
-    }
-
-    #[test]
-    fn the_cursor_is_always_inside_the_viewport() {
-        // The invariant the renderer depends on: `selected` must be one of the
-        // rows actually drawn — whatever the previous offset was, and however
-        // absurd the configured margin is.
-        for margin in [0usize, 3, 99] {
-            for total in [1usize, 5, 40] {
-                for cap in 1..12usize {
-                    for prev in 0..total {
-                        for sel in 0..total {
-                            let off = scroll_offset(prev, sel, cap, total, margin);
-                            let shown = cap.min(total);
-                            assert!(off <= total.saturating_sub(cap), "{margin}/{total}/{cap}");
-                            assert!(
-                                sel >= off && sel < off + shown,
-                                "{margin}/{total}/{cap}/{prev}/{sel}"
-                            );
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // --------------------------------------------------------- render_loading
-
-    /// The rendered rows, trailing blanks trimmed, paired with their y.
-    fn loading_rows(w: u16, h: u16) -> Vec<(u16, String)> {
-        use ratatui::backend::TestBackend;
-        let mut term = Terminal::new(TestBackend::new(w, h)).expect("test terminal");
-        term.draw(|f| render_loading(f, "connecting to Spotify", 0))
-            .expect("draw");
-        let buf = term.backend().buffer().clone();
-        (0..h)
-            .map(|y| {
-                let line: String = (0..w).map(|x| buf[(x, y)].symbol()).collect();
-                (y, line)
-            })
-            .filter(|(_, line)| !line.trim().is_empty())
-            .collect()
-    }
-
-    #[test]
-    fn the_loading_screen_names_what_it_is_waiting_on() {
-        let text = loading_rows(40, 12)
-            .into_iter()
-            .map(|(_, l)| l)
-            .collect::<String>();
-        assert!(text.contains("connecting to Spotify"), "{text}");
-        assert!(text.contains(SPINNER[0]), "spinner missing");
-        assert!(text.contains('\u{FF2D}'), "wordmark missing");
-    }
-
-    #[test]
-    fn the_loading_screen_is_centred_on_both_axes() {
-        let (w, h) = (41u16, 13u16);
-        let rows = loading_rows(w, h);
-        assert_eq!(rows.len(), 2, "expected a wordmark row and a spinner row");
-        for (y, line) in &rows {
-            // Measured from the content's midpoint, not its margins: the
-            // fullwidth wordmark leaves a blank continuation cell per letter.
-            let first = line.chars().take_while(|c| *c == ' ').count();
-            let last = line.trim_end().chars().count();
-            let centre = (first + last) / 2;
-            assert!(
-                centre.abs_diff(w as usize / 2) <= 1,
-                "row {y} centres at {centre}, screen centre {} ({line:?})",
-                w / 2
-            );
-        }
-        // Two rows of content, so the block straddles the middle of the screen.
-        let mid = (rows[0].0 + rows[1].0) / 2;
-        assert!(
-            mid.abs_diff(h / 2) <= 1,
-            "block sits at {mid}, screen mid {}",
-            h / 2
-        );
-    }
-
-    // ------------------------------------------------------------ retry_delay
-
-    #[test]
-    fn short_backoffs_are_waited_out_long_ones_are_not() {
-        assert!(retry_delay(Some(1)).is_some());
-        assert!(retry_delay(Some(5)).is_some());
-        assert!(
-            retry_delay(None).is_some(),
-            "no header means a token bounce"
-        );
-        // Spotify answers with hours once a development-mode quota is spent;
-        // sleeping on that froze a drill-in for minutes and failed anyway.
-        assert!(retry_delay(Some(60)).is_none());
-        assert!(retry_delay(Some(76_854)).is_none());
-    }
-
-    // ---------------------------------------------------------- scrub_target
-
-    #[test]
-    fn a_scrub_step_moves_by_the_step_size() {
-        assert_eq!(scrub_target(10_000, 200_000, 5_000), 15_000);
-        assert_eq!(scrub_target(10_000, 200_000, -5_000), 5_000);
-    }
-
-    #[test]
-    fn a_scrub_cannot_leave_the_track() {
-        // Held keys walk into both walls; neither may underflow or overshoot.
-        assert_eq!(scrub_target(2_000, 200_000, -5_000), 0);
-        assert_eq!(scrub_target(198_000, 200_000, 5_000), 200_000);
-        assert_eq!(scrub_target(0, 0, -5_000), 0);
-    }
-
-    // -------------------------------------------------------- drives_library
-
-    #[test]
-    fn zen_ignores_the_keys_that_only_move_the_hidden_library() {
-        for code in [
-            KeyCode::Tab,
-            KeyCode::BackTab,
-            KeyCode::Up,
-            KeyCode::Down,
-            KeyCode::Enter,
-            KeyCode::Char('j'),
-            KeyCode::Char('/'),
-            KeyCode::Char('o'),
-        ] {
-            assert!(drives_library(code), "{code:?} only drives the library");
-        }
-    }
-
-    #[test]
-    fn zen_keeps_playback_and_the_right_hand_pane() {
-        for code in [
-            KeyCode::Char(' '),
-            KeyCode::Char('n'),
-            KeyCode::Char('b'),
-            KeyCode::Char('s'),
-            KeyCode::Char('z'),
-            KeyCode::Char('q'),
-            KeyCode::Left,
-            KeyCode::Right,
-            // Retargets onto the playing track rather than going quiet.
-            KeyCode::Char('a'),
-        ] {
-            assert!(!drives_library(code), "{code:?} must still work in zen");
-        }
-    }
-
-    // ---------------------------------------------------------- ArtRepaint
-
-    #[test]
-    fn a_forced_repaint_blanks_the_box_before_redrawing_it() {
-        // The wipe is the whole mechanism: re-encoding is byte-identical for
-        // sixel and iTerm2, so a blank frame is the only thing the diff will
-        // emit — and it takes the overlay's leftovers with it.
-        assert_eq!(ArtRepaint::Wipe.advance(), ArtRepaint::Draw);
-        assert_eq!(ArtRepaint::Draw.advance(), ArtRepaint::Idle);
-        assert_eq!(ArtRepaint::Idle.advance(), ArtRepaint::Idle);
-    }
-
-    // ----------------------------------------------------------- should_draw
-
-    #[test]
-    fn input_redraws_at_the_next_terminal_refresh() {
-        // What made a held arrow key scroll in jerks: the cursor had moved but
-        // the frame was held back until the idle tick came round.
-        assert!(should_draw(true, false, MIN_FRAME));
-        assert!(!should_draw(
-            true,
-            false,
-            MIN_FRAME - Duration::from_millis(1)
-        ));
-    }
-
-    #[test]
-    fn an_untouched_screen_redraws_rarely() {
-        assert!(!should_draw(false, false, ANIM_FRAME));
-        assert!(should_draw(false, false, IDLE_REDRAW));
-    }
-
-    #[test]
-    fn animation_redraws_between_the_two() {
-        assert!(should_draw(false, true, ANIM_FRAME));
-        assert!(!should_draw(false, true, MIN_FRAME));
-    }
-
-    #[test]
-    fn a_fade_is_long_enough_to_be_smooth_at_the_animation_rate() {
-        // Smoothness has to come from duration, not frame rate: every present
-        // recomposes the viewport and the inline cover shimmers at 60fps.
-        let steps = FADE_MS / ANIM_FRAME.as_millis() as u64;
-        assert!(steps >= 30, "only {steps} steps of recolour");
-    }
-}
-
-/// Live-API tests, `#[ignore]`d so `cargo test` stays offline:
-///
-///     cargo test --bin myx -- --ignored --nocapture
-///
-/// They catch Spotify changing an endpoint out from under the artist page,
-/// which is how the 403/400 pair that emptied it went unnoticed.
-#[cfg(test)]
-mod live_tests {
-    use super::*;
-
-    fn token() -> Option<String> {
-        let path = myx::home_dir()?.join(".cache/myx/webapi.json");
-        let v: serde_json::Value =
-            serde_json::from_str(&std::fs::read_to_string(path).ok()?).ok()?;
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .ok()?
-            .as_secs();
-        if now + 30 >= v["expires_at"].as_u64().unwrap_or(0) {
-            eprintln!("cached token expired — run `myx` once, then retry");
-            return None;
-        }
-        v["access_token"].as_str().map(str::to_string)
-    }
-
-    /// girl in red: a real artist with a discography larger than one page.
-    const ARTIST_ID: &str = "3uwAm6vQy7kWPS2bciKWx9";
-    const ARTIST_NAME: &str = "girl in red";
-
-    /// The February 2026 library endpoints: check, save, check again. Uses a
-    /// track that is already saved, so a pass leaves the library as it was.
-    #[test]
-    #[ignore = "hits the Spotify API"]
-    fn live_library_contains_and_write() {
-        let Some(token) = token() else { return };
-        let client = http_client();
-        let v = get_json(&client, &format!("{API}/me/tracks?limit=1"), &token).expect("liked page");
-        let Some(id) = v["items"][0]["track"]["id"].as_str() else {
-            return println!("no liked tracks to test with");
-        };
-        let uri = format!("spotify:track:{id}");
-        assert!(
-            api_contains(&token, &uri),
-            "a track from /me/tracks reads as unsaved"
-        );
-        library_write(&client, &token, false, &uri).expect("re-saving an already saved track");
-        assert!(api_contains(&token, &uri), "still saved after the write");
-    }
-
-    /// Reproduces "liked failed http 403": page the whole Liked library the way
-    /// `spawn_library_fetch` does and report the first page Spotify refuses.
-    #[test]
-    #[ignore = "hits the Spotify API"]
-    fn live_liked_songs_page_all_the_way_through() {
-        let Some(token) = token() else { return };
-        let client = http_client();
-        let mut url = Some(format!("{API}/me/tracks?limit=50"));
-        let mut pages = 0;
-        let mut tracks = 0;
-        let mut total = None;
-        while let Some(u) = url.take() {
-            let resp = client.get(&u).bearer_auth(&token).send().expect("send");
-            let status = resp.status().as_u16();
-            let v: serde_json::Value = resp.json().unwrap_or(serde_json::Value::Null);
-            assert!(
-                (200..300).contains(&status),
-                "page {pages} -> HTTP {status}: {} ({u})",
-                v["error"]["message"].as_str().unwrap_or("(no message)")
-            );
-            total.get_or_insert(v["total"].as_u64().unwrap_or(0));
-            tracks += v["items"].as_array().map(Vec::len).unwrap_or(0);
-            url = v["next"].as_str().map(String::from);
-            pages += 1;
-        }
-        println!("liked: {tracks} tracks over {pages} pages, total={total:?}");
-        assert_eq!(Some(tracks as u64), total, "pagination dropped tracks");
-    }
-
-    #[test]
-    #[ignore = "hits the Spotify API"]
-    fn live_artist_top_tracks_are_not_empty() {
-        let Some(token) = token() else { return };
-        let rows = artist_top_tracks(&http_client(), &token, ARTIST_ID, ARTIST_NAME);
-        println!("top tracks: {}", rows.len());
-        for r in rows.iter().take(3) {
-            println!("  {} · {}", r.name, r.subtitle);
-        }
-        assert!(!rows.is_empty(), "the Popular section came back empty");
-        assert!(rows.iter().all(|r| r.is_track));
-    }
-
-    #[test]
-    #[ignore = "hits the Spotify API"]
-    fn live_artist_albums_page_past_the_first_ten() {
-        let Some(token) = token() else { return };
-        let rows = artist_albums(&http_client(), &token, ARTIST_ID);
-        println!("albums: {}", rows.len());
-        for r in rows.iter().take(3) {
-            println!("  {} · {}", r.name, r.subtitle);
-        }
-        if rows.is_empty() {
-            // A spent quota answers 429 for hours; that is not a code failure.
-            return println!("no albums came back — quota or API error");
-        }
-        // The bug was a single page rejected outright; more than one page's
-        // worth proves both that it succeeds and that `next` is followed.
-        assert!(
-            rows.len() > 10,
-            "expected a paged discography, got {}",
-            rows.len()
-        );
-        assert!(rows.iter().all(|r| !r.is_track && !r.is_header));
-        // Newest first.
-        let years: Vec<&str> = rows.iter().map(|r| r.subtitle.as_str()).collect();
-        assert!(
-            years.windows(2).all(|w| w[0] >= w[1]),
-            "not newest-first: {years:?}"
-        );
-    }
-
-    #[test]
-    #[ignore = "hits the Spotify API"]
-    fn live_artist_detail_has_both_sections() {
-        let Some(token) = token() else { return };
-        let (title, items) =
-            fetch_detail_blocking(&token, &format!("spotify:artist:{ARTIST_ID}"), ARTIST_NAME);
-        let headers: Vec<&str> = items
-            .iter()
-            .filter(|i| i.is_header)
-            .map(|i| i.name.as_str())
-            .collect();
-        println!("{title}: {} rows, headers {headers:?}", items.len());
-        assert!(
-            headers.contains(&"Popular"),
-            "artist page lost its Popular section: {headers:?}"
-        );
-        if !headers.contains(&"Albums") {
-            // A spent album-endpoint quota answers 429 for hours, and the
-            // section drops out; that is not a code failure.
-            return println!("no Albums section — quota or API error");
-        }
-        assert_eq!(headers, ["Popular", "Albums"]);
-        assert!(items.len() > 20, "only {} rows", items.len());
-    }
-}
+#[path = "main_tests/mod.rs"]
+mod main_tests;
